Sf4.3 compatibility
